feat: adiciona quantidade minima e maxima para produto e implementa notificacao de status de estoque

+++ backend/src/main/java/com/hextech/estoque_api/domain/services/StockMovementDomainService.java
@@ -1,15 +1,17 @@
 package com.hextech.estoque_api.domain.services;
 
 import com.hextech.estoque_api.domain.entities.*;
+import com.hextech.estoque_api.domain.entities.product.Product;
 import com.hextech.estoque_api.domain.exceptions.InvalidMovementTypeException;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Service
 public class StockMovementDomainService {
 
-    public Movement processMovement(MovementType type, Integer quantity, String note,
+    public Movement processMovement(MovementType type, BigDecimal quantity, String note,
                                     Product product, User user, Company company, StockLocation stockLocation) {
         if (type.equals(MovementType.ENTRADA)) {
             product.increaseQuantity(quantity);
